+++ src/modules/components/breadcrumb/index.tsx
@@ -1,0 +1,1 @@
+export { Breadcrumb, type IBreadcrumbProps } from './breadcrumb';