+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `AssetTransfer` module component
+-   Implement `AssetTransfer` and `Breadcrumb` module components
+-   Added `slash` icon file
 
 ### Changed